+++ actxps/expose.py
@@ -6,9 +6,10 @@ from datetime import date
 from actxps.tools import (
     arg_match,
     _verify_col_names,
-    _check_convert_df
+    _check_convert_df,
+    _check_missing_dates
 )
-from actxps.dates import frac_interval, _date_str
+from actxps.dates import _delta_frac, _date_str
 from warnings import warn
 from functools import singledispatchmethod
 from itertools import product
@@ -18,7 +19,7 @@ class ExposedDF():
     """
     Exposed data frame class
 
-    Convert a data frame of census-level records into an object with 
+    Convert a data frame of census-level records into an object with
     exposure-level records.
 
 
@@ -27,10 +28,10 @@ class ExposedDF():
     data : pl.DataFrame | pd.DataFrame
         A data frame with census-level records
     end_date : date | str
-        Experience study end date. If a string is passed, it must be in 
+        Experience study end date. If a string is passed, it must be in
         %Y-%m-%d format.
     start_date : date | str, default=date(1900, 1, 1)
-        Experience study start date. If a string is passed, it must be in 
+        Experience study start date. If a string is passed, it must be in
         %Y-%m-%d format.
     target_status : str | list | np.ndarray, default=`None`
         Target status values
@@ -59,7 +60,8 @@ class ExposedDF():
         existing columns in the original input data plus new columns for
         exposures and observation periods. Observation periods include counters
         for policy exposures, start dates, and end dates. Both start dates and
-        end dates are inclusive bounds.
+        end dates are inclusive bounds. All columns containing dates must be in
+        YYYY-MM-DD format.
 
         For policy year exposures, two observation period columns are returned.
         Columns beginning with (`pol_`) are integer policy periods. Columns
@@ -74,7 +76,7 @@ class ExposedDF():
         Names of the start and end date columns in `data` for each exposure
         period
     trx_types: list
-        List of transaction types that have been attached to `data` using 
+        List of transaction types that have been attached to `data` using
         the `add_transactions()` method.
 
 
@@ -102,7 +104,7 @@ class ExposedDF():
 
     **Alternative class constructors**
 
-    - `expose_py()`, `expose_pq()`, `expose_pm()`, `expose_pw()`, `expose_cy()`, 
+    - `expose_py()`, `expose_pq()`, `expose_pm()`, `expose_pw()`, `expose_cy()`,
         `expose_cq()`, `expose_cm()`, `expose_cw()`
 
         Convenience constructor functions for specific exposure calculations.
@@ -116,11 +118,11 @@ class ExposedDF():
         `q` = quarters
         `m` = months
         `w` = weeks
-        Each constructor has the same inputs as the `__init__` method except 
+        Each constructor has the same inputs as the `__init__` method except
         that `expo_length` and `cal_expo` arguments are prepopulated.
 
     - `from_DataFrame()`
-        Convert a data frame that already has exposure-level records into an 
+        Convert a data frame that already has exposure-level records into an
         `ExposedDF` object.
 
 
@@ -135,7 +137,7 @@ class ExposedDF():
     ```{python}
     import actxps as xp
 
-    xp.ExposedDF(xp.load_toy_census(), "2020-12-31", 
+    xp.ExposedDF(xp.load_toy_census(), "2020-12-31",
                  target_status='Surrender')
     ```
     """
@@ -189,19 +191,15 @@ class ExposedDF():
         abbrev = ExposedDF.abbr_period[expo_length]
         interval = ExposedDF.abbr_pl[expo_length]
 
-        # call frac_intervals using a polars struct
-        def per_frac(x):
-            return frac_interval(x.struct[0], x.struct[1], expo_length)
+        # call _delta_frac using a polars struct
+        def per_frac(x, cal=False):
+            return _delta_frac(x.struct[0], x.struct[1], expo_length, cal)
 
         # add time intervals to a date expression
         def add_per(x: pl.Expr, n: pl.Expr):
             return (x.dt.offset_by(pl.format('{}', n.cast(str) + interval)))
 
-        if cal_expo:
-            if expo_length != 'week':
-                floor_date = '1' + interval,
-            else:
-                floor_date = ('1' + interval, '-1d')
+        n_term_dates = data[col_term_date].is_not_null().sum()
 
         # column renames and name conflicts
         data = data.rename({
@@ -209,7 +207,15 @@ class ExposedDF():
             col_status: 'status',
             col_issue_date: 'issue_date',
             col_term_date: 'term_date'
-        })
+        }).with_columns(
+            pl.col("issue_date", "term_date").cast(pl.Date)
+        )
+
+        _check_missing_dates(data["issue_date"])
+
+        assert n_term_dates == data["term_date"].is_not_null().sum(), \
+            ("Bad termination date formats were detected.\n" +
+             "Make sure all dates are in YYYY-MM-DD format.")
 
         # check for potential name conflicts
         x = {"exposure",
@@ -238,6 +244,9 @@ class ExposedDF():
             default_status = pl.Series([default_status],
                                        dtype=status_levels)
 
+        assert default_status[0] not in target_status, \
+            "`default_status` is not allowed to be the same as `target_status`"
+
         # pre-exposure updates
         # drop policies issued after the study end and
         #   policies that terminated before the study start
@@ -258,16 +267,22 @@ class ExposedDF():
 
         if cal_expo:
 
+            if expo_length != 'week':
+                off = ''
+            else:
+                off = '-1d'
+
             start_dates = pl.Series(pl.repeat(start_date, len(data),
                                               eager=True))
             data = data.with_columns(
                 first_date=pl.max_horizontal('issue_date', start_dates),
             ).with_columns(
-                cal_b=pl.col('first_date').dt.truncate(*floor_date)
+                cal_b=(pl.col('first_date').dt.truncate('1' + interval).
+                       dt.offset_by(off))
             ).with_columns(
                 tot_per=pl.struct(pl.col('cal_b').dt.offset_by('-1d'),
                                   pl.col('last_date')).
-                map_batches(per_frac)
+                map_batches(lambda x: per_frac(x, True))
             )
 
         else:
@@ -895,7 +910,8 @@ class ExposedDF():
         trx_data : pl.DataFrame | pd.DataFrame
             A data frame containing transactions details. This data frame must
             have columns for policy numbers, transaction dates, transaction
-            types, and transaction amounts.
+            types, and transaction amounts. Transaction dates must be in 
+            YYYY-MM-DD format.
         col_pol_num : str, default='pol_num'
             Name of the column in `trx_data` containing the policy number
         col_trx_date : str, default='trx_date'
@@ -943,7 +959,11 @@ class ExposedDF():
             col_trx_date: 'trx_date',
             col_trx_type: 'trx_type',
             col_trx_amt: 'trx_amt'
-        })
+        }).with_columns(
+            pl.col("trx_date").cast(pl.Date)
+        )
+
+        _check_missing_dates(trx_data["trx_date"])
 
         # check for conflicting transaction types
         new_trx_types = trx_data['trx_type'].unique().to_list()